feat: stats-section-underlines rename files to fit convention
