+++ src/app/_main-page-components/HeroSection.tsx
