+++ src/app/_main-page-components/MenopauseStagesSection.tsx
