+++ src/app/_main-page-components/stats-section/StatsSection.tsx
